+++ website/consent-default.js
@@ -1,0 +1,16 @@
+window.dataLayer = window.dataLayer || []
+
+function gtag() {
+  window.dataLayer.push(arguments)
+}
+
+gtag('consent', 'default', {
+  ad_storage: 'denied',
+  analytics_storage: 'denied',
+  ad_user_data: 'denied',
+  ad_personalization: 'denied',
+  functionality_storage: 'granted',
+  personalization_storage: 'denied',
+  security_storage: 'granted',
+  wait_for_update: 500,
+})