+++ website/ga4-init.js
@@ -1,0 +1,5 @@
+gtag('js', new Date())
+gtag('config', 'G-C1E2XBE47T', {
+  allow_google_signals: false,
+  allow_ad_personalization_signals: false,
+})